Logging fully implemented EventRequest is now an EventEmitter
need to fix the formatting

+++ server/middlewares/body_parsers/multipart_data_parser.js
@@ -632,7 +632,7 @@ class MultipartFormParser extends BodyParser
 			this.callback( false, this.parts );
 		});
 
-		this.event.getEventEmitter().on( 'cleanUp', () => {
+		this.event.on( 'cleanUp', () => {
 			this.cleanUpItems();
 			this.terminate();
 		});
@@ -646,7 +646,7 @@ class MultipartFormParser extends BodyParser
 	cleanUpItems()
 	{
 		this.parts.files.forEach( ( part ) =>{
-			if ( part.type === DATA_TYPE_FILE && part.path !== 'undefined' )
+			if ( part.type === DATA_TYPE_FILE && part.path !== 'undefined' && fs.existsSync( part.path ) )
 			{
 				fs.unlinkSync( part.path )
 			}
